requirements.txt
faker
pandas
psycopg2-binary
sqlalchemy
streamlit